requirements copy.txt
pillow==8.4.0
numpy==1.19.5
streamlit==1.4.0
tensorflow==2.4.0
